viewer_app/requirements.txt
Flask>=3.0.0
google-cloud-firestore
gunicorn
